refactor(core): split metadata state into rows, apply, and queries
