+++ crates/loonfs-core/src/metadata/rows.rs
@@ -1,0 +1,388 @@
+//! Metadata row records plus the append and accounting plumbing that keeps
+//! [`MetadataState`]'s derived indexes and decoded-size totals in step with
+//! its append-only rows.
+
+use super::indexes::MetadataIndexes;
+use loonfs_api::{ChangeSeq, CommitId, ContentRef, InodeId, InodeKind, NameKey, RevisionNo};
+use serde::{Deserialize, Serialize};
+use std::mem::{size_of, size_of_val};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
+pub struct MetadataState {
+    #[serde(default)]
+    pub(super) inodes: Vec<InodeRecord>,
+    #[serde(default)]
+    pub(super) direntry_binds: Vec<DirentryBindRecord>,
+    #[serde(default)]
+    pub(super) direntry_unbinds: Vec<DirentryUnbindRecord>,
+    #[serde(default)]
+    pub(super) revisions: Vec<RevisionRecord>,
+    #[serde(default)]
+    pub(super) subtree_tombstones: Vec<SubtreeTombstoneRecord>,
+    #[serde(default)]
+    pub(super) commit_receipts: Vec<CommitReceiptRecord>,
+    #[serde(skip)]
+    pub(super) row_count: usize,
+    #[serde(skip)]
+    pub(super) decoded_bytes: usize,
+    #[serde(skip)]
+    pub(super) indexes: MetadataIndexes,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Default, Deserialize)]
+struct MetadataStateRows {
+    #[serde(default)]
+    inodes: Vec<InodeRecord>,
+    #[serde(default)]
+    direntry_binds: Vec<DirentryBindRecord>,
+    #[serde(default)]
+    direntry_unbinds: Vec<DirentryUnbindRecord>,
+    #[serde(default)]
+    revisions: Vec<RevisionRecord>,
+    #[serde(default)]
+    subtree_tombstones: Vec<SubtreeTombstoneRecord>,
+    #[serde(default)]
+    commit_receipts: Vec<CommitReceiptRecord>,
+}
+
+impl Default for MetadataState {
+    fn default() -> Self {
+        Self::from_rows(
+            Vec::new(),
+            Vec::new(),
+            Vec::new(),
+            Vec::new(),
+            Vec::new(),
+            Vec::new(),
+        )
+    }
+}
+
+impl<'de> Deserialize<'de> for MetadataState {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let rows = MetadataStateRows::deserialize(deserializer)?;
+        Ok(Self::from_rows(
+            rows.inodes,
+            rows.direntry_binds,
+            rows.direntry_unbinds,
+            rows.revisions,
+            rows.subtree_tombstones,
+            rows.commit_receipts,
+        ))
+    }
+}
+
+/// Converts a metadata record name key back to the wire [`NameKey`] type.
+///
+/// Record name keys only enter [`MetadataState`] from validated sources:
+/// typed manifest rows, commit plans built from typed [`NameKey`]s, and WAL
+/// deltas whose name keys replay validation checks before applying. An
+/// invalid key here is therefore a logic bug, not untrusted input.
+pub(crate) fn record_name_key(name_key: &str) -> NameKey {
+    NameKey::parse(name_key).expect("metadata record name keys are validated on ingress")
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct InodeRecord {
+    pub inode_id: InodeId,
+    pub inode_kind: InodeKind,
+    pub created_seq: ChangeSeq,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct DirentryBindRecord {
+    pub parent_inode_id: InodeId,
+    pub name_key: String,
+    pub display_name: String,
+    pub child_inode_id: InodeId,
+    pub bind_seq: ChangeSeq,
+    pub bind_delta_index: u32,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct DirentryUnbindRecord {
+    pub parent_inode_id: InodeId,
+    pub name_key: String,
+    pub child_inode_id: InodeId,
+    pub bind_seq: ChangeSeq,
+    pub bind_delta_index: u32,
+    pub unbind_seq: ChangeSeq,
+    pub unbind_delta_index: u32,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct RevisionRecord {
+    pub inode_id: InodeId,
+    pub revision_no: RevisionNo,
+    pub committed_seq: ChangeSeq,
+    pub revision_delta_index: u32,
+    pub content_ref: ContentRef,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct SubtreeTombstoneRecord {
+    pub root_inode_id: InodeId,
+    pub tombstone_seq: ChangeSeq,
+    pub tombstone_delta_index: u32,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct CommitReceiptRecord {
+    pub commit_id: CommitId,
+    pub semantic_commit_fingerprint: String,
+    pub committed_seq: ChangeSeq,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub message: Option<String>,
+}
+
+impl MetadataState {
+    pub(crate) fn from_rows(
+        inodes: Vec<InodeRecord>,
+        direntry_binds: Vec<DirentryBindRecord>,
+        direntry_unbinds: Vec<DirentryUnbindRecord>,
+        revisions: Vec<RevisionRecord>,
+        subtree_tombstones: Vec<SubtreeTombstoneRecord>,
+        commit_receipts: Vec<CommitReceiptRecord>,
+    ) -> Self {
+        let mut state = Self {
+            inodes,
+            direntry_binds,
+            direntry_unbinds,
+            revisions,
+            subtree_tombstones,
+            commit_receipts,
+            row_count: 0,
+            decoded_bytes: 0,
+            indexes: MetadataIndexes::default(),
+        };
+        state.rebuild_indexes();
+        state
+    }
+
+    /// Highest sequence carried by any indexed record.
+    ///
+    /// No record carries a larger seq, so a query at `base_seq >=
+    /// indexed_seq()` passes every `seq <= base_seq` filter and is equivalent
+    /// to an at-head query. The seq-gated read methods below rely on this to
+    /// route to the indexes; only queries strictly below `indexed_seq()` need
+    /// the historical scans.
+    pub fn indexed_seq(&self) -> ChangeSeq {
+        self.indexes.indexed_seq()
+    }
+
+    pub fn inodes(&self) -> &[InodeRecord] {
+        &self.inodes
+    }
+
+    pub fn direntry_binds(&self) -> &[DirentryBindRecord] {
+        &self.direntry_binds
+    }
+
+    pub fn direntry_unbinds(&self) -> &[DirentryUnbindRecord] {
+        &self.direntry_unbinds
+    }
+
+    pub fn revisions(&self) -> &[RevisionRecord] {
+        &self.revisions
+    }
+
+    pub fn subtree_tombstones(&self) -> &[SubtreeTombstoneRecord] {
+        &self.subtree_tombstones
+    }
+
+    pub fn commit_receipts(&self) -> &[CommitReceiptRecord] {
+        &self.commit_receipts
+    }
+
+    pub fn row_count(&self) -> usize {
+        self.row_count
+    }
+
+    pub fn decoded_bytes(&self) -> usize {
+        self.decoded_bytes
+    }
+
+    pub fn find_commit_receipt(&self, commit_id: &CommitId) -> Option<&CommitReceiptRecord> {
+        self.indexes.commit_receipt(commit_id)
+    }
+
+    fn rebuild_indexes(&mut self) {
+        self.row_count = metadata_row_count(
+            &self.inodes,
+            &self.direntry_binds,
+            &self.direntry_unbinds,
+            &self.revisions,
+            &self.subtree_tombstones,
+            &self.commit_receipts,
+        );
+        self.decoded_bytes = metadata_decoded_bytes(
+            &self.inodes,
+            &self.direntry_binds,
+            &self.direntry_unbinds,
+            &self.revisions,
+            &self.subtree_tombstones,
+            &self.commit_receipts,
+        );
+        self.indexes = MetadataIndexes::rebuild(
+            &self.inodes,
+            &self.direntry_binds,
+            &self.direntry_unbinds,
+            &self.revisions,
+            &self.subtree_tombstones,
+            &self.commit_receipts,
+        );
+    }
+
+    pub(crate) fn push_inode_record(&mut self, record: InodeRecord) {
+        self.indexes.record_inode(&record);
+        self.record_row_weight(size_of::<InodeRecord>());
+        self.inodes.push(record);
+    }
+
+    pub(crate) fn push_direntry_bind_record(&mut self, record: DirentryBindRecord) {
+        self.indexes.record_bind(&record);
+        self.record_row_weight(direntry_bind_decoded_bytes(&record));
+        self.direntry_binds.push(record);
+    }
+
+    pub(crate) fn push_direntry_unbind_record(&mut self, record: DirentryUnbindRecord) {
+        self.indexes.record_unbind(&record);
+        self.record_row_weight(direntry_unbind_decoded_bytes(&record));
+        self.direntry_unbinds.push(record);
+    }
+
+    pub(crate) fn push_revision_record(&mut self, record: RevisionRecord) {
+        self.indexes.record_revision(&record);
+        self.record_row_weight(revision_decoded_bytes(&record));
+        self.revisions.push(record);
+    }
+
+    pub(crate) fn push_subtree_tombstone_record(&mut self, record: SubtreeTombstoneRecord) {
+        self.indexes.record_tombstone(&record);
+        self.record_row_weight(size_of::<SubtreeTombstoneRecord>());
+        self.subtree_tombstones.push(record);
+    }
+
+    pub(crate) fn push_commit_receipt_record(&mut self, record: CommitReceiptRecord) {
+        self.indexes.record_commit_receipt(&record);
+        self.record_row_weight(commit_receipt_decoded_bytes(&record));
+        self.commit_receipts.push(record);
+    }
+
+    fn record_row_weight(&mut self, decoded_bytes: usize) {
+        self.row_count = self.row_count.saturating_add(1);
+        self.decoded_bytes = self.decoded_bytes.saturating_add(decoded_bytes);
+    }
+}
+
+#[cfg(test)]
+#[derive(Debug, Default)]
+pub(crate) struct MetadataStateBuilder {
+    state: MetadataState,
+}
+
+#[cfg(test)]
+impl MetadataStateBuilder {
+    pub(crate) fn push_inode(&mut self, record: InodeRecord) {
+        self.state.push_inode_record(record);
+    }
+
+    pub(crate) fn push_direntry_bind(&mut self, record: DirentryBindRecord) {
+        self.state.push_direntry_bind_record(record);
+    }
+
+    pub(crate) fn push_direntry_unbind(&mut self, record: DirentryUnbindRecord) {
+        self.state.push_direntry_unbind_record(record);
+    }
+
+    pub(crate) fn push_revision(&mut self, record: RevisionRecord) {
+        self.state.push_revision_record(record);
+    }
+
+    pub(crate) fn push_subtree_tombstone(&mut self, record: SubtreeTombstoneRecord) {
+        self.state.push_subtree_tombstone_record(record);
+    }
+
+    pub(crate) fn push_commit_receipt(&mut self, record: CommitReceiptRecord) {
+        self.state.push_commit_receipt_record(record);
+    }
+
+    pub(crate) fn finish(mut self) -> MetadataState {
+        self.state.rebuild_indexes();
+        self.state
+    }
+}
+
+fn metadata_row_count(
+    inodes: &[InodeRecord],
+    direntry_binds: &[DirentryBindRecord],
+    direntry_unbinds: &[DirentryUnbindRecord],
+    revisions: &[RevisionRecord],
+    subtree_tombstones: &[SubtreeTombstoneRecord],
+    commit_receipts: &[CommitReceiptRecord],
+) -> usize {
+    inodes
+        .len()
+        .saturating_add(direntry_binds.len())
+        .saturating_add(direntry_unbinds.len())
+        .saturating_add(revisions.len())
+        .saturating_add(subtree_tombstones.len())
+        .saturating_add(commit_receipts.len())
+}
+
+fn metadata_decoded_bytes(
+    inodes: &[InodeRecord],
+    direntry_binds: &[DirentryBindRecord],
+    direntry_unbinds: &[DirentryUnbindRecord],
+    revisions: &[RevisionRecord],
+    subtree_tombstones: &[SubtreeTombstoneRecord],
+    commit_receipts: &[CommitReceiptRecord],
+) -> usize {
+    size_of_val(inodes)
+        .saturating_add(
+            direntry_binds
+                .iter()
+                .map(direntry_bind_decoded_bytes)
+                .sum::<usize>(),
+        )
+        .saturating_add(
+            direntry_unbinds
+                .iter()
+                .map(direntry_unbind_decoded_bytes)
+                .sum::<usize>(),
+        )
+        .saturating_add(revisions.iter().map(revision_decoded_bytes).sum::<usize>())
+        .saturating_add(size_of_val(subtree_tombstones))
+        .saturating_add(
+            commit_receipts
+                .iter()
+                .map(commit_receipt_decoded_bytes)
+                .sum::<usize>(),
+        )
+}
+
+fn direntry_bind_decoded_bytes(record: &DirentryBindRecord) -> usize {
+    size_of::<DirentryBindRecord>() + record.name_key.len() + record.display_name.len()
+}
+
+fn direntry_unbind_decoded_bytes(record: &DirentryUnbindRecord) -> usize {
+    size_of::<DirentryUnbindRecord>() + record.name_key.len()
+}
+
+fn revision_decoded_bytes(record: &RevisionRecord) -> usize {
+    size_of::<RevisionRecord>() + content_ref_decoded_bytes(&record.content_ref)
+}
+
+fn commit_receipt_decoded_bytes(record: &CommitReceiptRecord) -> usize {
+    size_of::<CommitReceiptRecord>()
+        + record.commit_id.as_str().len()
+        + record.semantic_commit_fingerprint.len()
+        + record.message.as_ref().map_or(0, String::len)
+}
+
+fn content_ref_decoded_bytes(content_ref: &ContentRef) -> usize {
+    size_of::<ContentRef>() + content_ref.digest.len()
+}
